#int -> double

+++ src/main/java/cn/edu/nju/modules/schedule/SchedulerImpl.java
@@ -16,8 +16,8 @@ public class SchedulerImpl implements Scheduler {
     public void schedule() {
         // 遍历所有种子，调整其能量
         for (Seed seed : resourcesManager.getAllSeeds()) {
-            int currentEnergy = seed.getEnergy();
-            int newCoverage = seed.getCoverage();
+            double currentEnergy = seed.getEnergy();
+            double newCoverage = seed.getCoverage();
 
             // 简单的能量调度策略：覆盖率越高，能量越高
             if (newCoverage > currentEnergy) {
